+++ src/media/encoders/media.encoder.wav.cpp
@@ -74,7 +74,7 @@ public:
 	virtual dseed::error_t commit () override
 	{
 		_stream->seek (dseed::seekorigin_begin, 4);
-		uint32_t totalSize = _stream->length ();
+		uint32_t totalSize = (uint32_t)_stream->length ();
 		_stream->write (&totalSize, 4);
 		uint32_t dataSeek = sizeof (RIFF_HEADER) + sizeof (WAVE_HEADER) + sizeof (fmt_HEADER) + 4;
 		_stream->seek (dseed::seekorigin_begin, dataSeek);
